==> notebooks_&_data/1a.preprocess_columns.ipynb <==
# Data preprocessing

This series of notebooks outline the steps undertook to preprocess the sampling data in preparation for EDA, label generation, and modeling.

Step 1: Clean up the columns

# Import libraries

import numpy as np
import pandas as pd

# Load the data

df = pd.read_csv("inputs/random_opinion.csv")
df

# Clean-up the columns

### remove columns that don't have any values

print("number of columns: ", len(df.columns))
df = df.dropna(axis=1, how="all")
print("number of remaining columns: ", len(df.columns))

### clean-up the opinion columns

columns = ["id", "date_created", "date_modified", "type", "sha1", "download_url", "local_path",
    "plain_text", "html", "html_lawbox", "html_columbia", "html_with_citations",
    "extracted_by_ocr", "author_id", "cluster_id", "per_curiam", "page_count",
    "author_str", "joined_by_str", "xml_harvard", "html_anon_2020", "ordering_key"]
print("number of columns: ", len(columns))

dropped_columns = set(columns) - set(df.columns)
print("columns dropped: ", list(dropped_columns))
remaining_columns = set(columns) - set(dropped_columns)
print("number of columns remaining: ", len(remaining_columns))

columns_to_drop = ["date_modified", "sha1", "download_url", "local_path", "author_id", "author_str", "page_count"]
df = df.drop(columns=columns_to_drop)

print("columns dropped: ", list(columns_to_drop))
remaining_columns = set(remaining_columns) - set(columns_to_drop)
print("number of columns remaining: ", len(remaining_columns))

print("opinion columns: ", list(remaining_columns))

### clean-up the cluster columns

columns = ["id.1", "judges", "date_created.1", "date_modified.1", "date_filed", "slug",
    "case_name_short", "case_name", "case_name_full", "scdb_id", "source",
    "procedural_history", "attorneys", "nature_of_suit", "posture", "syllabus",
    "citation_count", "precedential_status", "date_blocked", "blocked", "docket_id",
    "scdb_decision_direction", "scdb_votes_majority", "scdb_votes_minority",
    "date_filed_is_approximate", "correction", "cross_reference", "disposition",
    "filepath_json_harvard", "headnotes", "history", "other_dates", "summary",
    "arguments", "headmatter", "filepath_pdf_harvard"]

dropped_columns = set(columns) - set(df.columns)
print("columns dropped: ", list(dropped_columns))
remaining_columns = set(columns) - set(dropped_columns)
print("number of columns remaining: ", len(remaining_columns))

columns_to_drop = ["id.1", "date_created.1", "date_modified.1", "date_filed", "slug", "scdb_id", "attorneys",
                  "date_blocked", "scdb_decision_direction", "scdb_votes_majority", "scdb_votes_minority",
                  "date_filed_is_approximate", "filepath_json_harvard", "other_dates"]
df = df.drop(columns=columns_to_drop)

print("columns dropped: ", list(columns_to_drop))
remaining_columns = set(remaining_columns) - set(columns_to_drop)
print("number of columns remaining: ", len(remaining_columns))

print("cluster columns: ", list(remaining_columns))

### clean-up the docket columns

columns = ["id.2", "date_created.2", "date_modified.2", "date_argued", "date_reargued",
    "date_reargument_denied", "case_name_short.1", "case_name.1", "case_name_full.1",
    "slug.1", "docket_number", "date_blocked.1", "blocked.1", "court_id",
    "date_cert_denied", "date_cert_granted", "assigned_to_id", "cause", "date_filed.1",
    "date_last_filing", "date_terminated", "filepath_ia", "filepath_local",
    "jurisdiction_type", "jury_demand", "nature_of_suit.1", "pacer_case_id",
    "referred_to_id", "source.1", "assigned_to_str", "referred_to_str", "view_count",
    "date_last_index", "appeal_from_id", "appeal_from_str", "appellate_case_type_information",
    "appellate_fee_status", "panel_str", "originating_court_information_id", "mdl_status",
    "filepath_ia_json", "ia_date_first_change", "ia_needs_upload", "ia_upload_failure_count",
    "docket_number_core", "idb_data_id", "federal_defendant_number", "federal_dn_case_type",
    "federal_dn_judge_initials_assigned", "federal_dn_judge_initials_referred",
    "federal_dn_office_code", "parent_docket_id"
]

dropped_columns = set(columns) - set(df.columns)
print("columns dropped: ", list(dropped_columns))
remaining_columns = set(columns) - set(dropped_columns)
print("number of columns remaining: ", len(remaining_columns))

columns_to_drop = ["id.2", "slug.1", 'blocked.1', 'case_name.1', 'case_name_full.1', 'case_name_short.1',
                  'date_blocked.1','date_created.2','date_filed.1', 'date_modified.2','nature_of_suit.1',
                  'source.1', 'idb_data_id', 'date_argued', 'appeal_from_id','appeal_from_str',
                   'appellate_case_type_information','appellate_fee_status','assigned_to_id','assigned_to_str',
                  'date_last_filing','date_last_index','date_reargument_denied','date_terminated', 
                   'filepath_ia_json','ia_date_first_change', 'ia_needs_upload','ia_upload_failure_count',
                   'originating_court_information_id', 'pacer_case_id', 'docket_number_core', "jurisdiction_type",
                  "referred_to_str"]
df = df.drop(columns=columns_to_drop)

print("columns dropped: ", list(columns_to_drop))
remaining_columns = set(remaining_columns) - set(columns_to_drop)
print("number of columns remaining: ", len(remaining_columns))

print("docket columns: ", list(remaining_columns))

### clean-up the court columns

columns = court_columns = ["id.3", "date_modified.3", "in_use", "has_opinion_scraper", "has_oral_argument_scraper",
    "position", "citation_string", "short_name", "full_name", "url", "start_date",
    "end_date", "jurisdiction", "notes", "pacer_court_id", "fjc_court_id",
    "pacer_has_rss_feed", "date_last_pacer_contact", "pacer_rss_entry_types",
    "parent_court_id"
]

dropped_columns = set(columns) - set(df.columns)
print("columns dropped: ", list(dropped_columns))
remaining_columns = set(columns) - set(dropped_columns)
print("number of columns remaining: ", len(remaining_columns))

columns_to_drop = ['id.3', 'date_modified.3', 'start_date', 'end_date', 'url','citation_string',
                  'position','fjc_court_id','has_opinion_scraper','has_oral_argument_scraper','notes',
                  'pacer_court_id','pacer_has_rss_feed','pacer_rss_entry_types','parent_court_id']
df = df.drop(columns=columns_to_drop)

print("columns dropped: ", list(columns_to_drop))
remaining_columns = set(remaining_columns) - set(columns_to_drop)
print("number of columns remaining: ", len(remaining_columns))

print("court columns: ", list(remaining_columns))

### rename & reorder all columns

opinion_columns = {"id": "opinion_id", 
                   "date_created": "opinion_date_created",
                   "type": "opinion_type",
                   "extracted_by_ocr": "opinion_extracted_by_ocr",
                   "per_curiam": "opinion_per_curiam",
                   "plain_text": "opinion_plain_text",
                   "html": "opinion_html",
                   "html_with_citations": "opinion_html_with_citations",
                   "html_anon_2020": "opinion_html_anon_2020",
                   "html_columbia": "opinion_html_columbia",
                   "html_lawbox": "opinion_html_lawbox",
                   "xml_harvard": "opinion_xml_harvard"}

cluster_columns = {"cluster_id": "cluster_id",
                   "judges": "cluster_judges",
                   "nature_of_suit": "cluster_nature_of_suit",
                   "source": "cluster_source",
                   "blocked": "cluster_blocked",
                   "precedential_status": "cluster_precedential_status",
                   "citation_count": "cluster_citation_count",
                   "case_name": "cluster_case_name",
                   "case_name_short": "cluster_case_name_short",
                   "case_name_full": "cluster_case_name_full",
                   "summary": "cluster_summary",
                   "history": "cluster_history",
                   "headmatter": "cluster_headmatter",
                   "headnotes": "cluster_headnotes",
                   "posture": "cluster_posture",
                   "arguments": "cluster_arguments",
                   "cross_reference": "cluster_cross_reference",
                   "disposition": "cluster_disposition",
                   "syllabus": "cluster_syllabus"}

docket_columns = {'docket_id': 'docket_id',
                  'docket_number': 'docket_number',
                  'view_count': 'docket_view_count'}

court_columns = {'court_id': 'court_id',
                 'jurisdiction': 'court_jurisdiction', 
                 'in_use': 'court_in_use',
                 'short_name': 'court_short_name',
                 'full_name': 'court_full_name'}

combined_columns = {}
for d in [opinion_columns, cluster_columns, docket_columns, court_columns]:
    combined_columns.update(d)

df.rename(columns=combined_columns, inplace=True)
df = df[list(combined_columns.values())]

df.head()

print("number of columns: ", len(df.columns))
df.columns

### save the sample data with cleaned columns

df.to_csv("outputs/1a.columns_cleaned.csv", index=False)
