==> notebooks_&_data/1b.preprocess_opinions.ipynb <==
# Data preprocessing

This series of notebooks outline the steps undertook to preprocess the sampling data in preparation for EDA, label generation, and modeling.

Step 2: Clean-up the opinions from various sources and combine them to one column

# Import libraries

import numpy as np
import pandas as pd

from bs4 import BeautifulSoup
from eyecite import clean_text

# Helper functions

# Function to recursively extract xml text elements and combine them into one string
def extract_text(tag):
    combined_text = '' 
    
    # Extract the text, stripping leading/trailing whitespace
    text = tag.get_text(strip=False)
    if text:
        return text
        
    # Recursively process all child elements (if any)
    for child in tag.find_all(True):  # True matches all tags
        combined_text += extract_text(child)  # Concatenate child text recursively
    return combined_text

# Function to clean opinion_text in xml format
def clean_xml(opinion_text):
    # Use BeautifulSoup to load the xml opinion_text
    soup = BeautifulSoup(opinion_text, 'lxml-xml')
    
    # Remove the text inside <a> tags
    for a_tag in soup.find_all('a'):
        a_tag.string = '' 

    # Extract the texts recursively over each element
    extracted_text = extract_text(soup)

    # Clean up the extracted text using eyecite
    cleaned_text = clean_text(extracted_text, ['all_whitespace', 'underscores'])
    return cleaned_text

# Function to clean opinion_text in html format
def clean_html(opinion_text):
    cleaned_text = clean_text(opinion_text, ['html', 'all_whitespace', 'underscores'])
    return cleaned_text

# Function to clean opinion_text in plain_text format
def clean_plain(opinion_text):
    cleaned_text = clean_text(opinion_text, ['all_whitespace', 'underscores'])
    return cleaned_text

# Function to get the first non-null value from the specified columns
def get_first_non_null(row, columns):
    for col in columns:
        if pd.notna(row[col]):
            return row[col], col
    return None, None  # Return None if all columns are null

# Load the data

df = pd.read_csv("outputs/1a.columns_cleaned.csv")
df

df.columns

### confirm opinion_id is the unique identifier for each opinion

assert df['opinion_id'].nunique() == len(df['opinion_id'])

opinions_columns = ['opinion_plain_text',
       'opinion_html', 'opinion_html_with_citations', 'opinion_html_anon_2020',
       'opinion_html_columbia', 'opinion_html_lawbox', 'opinion_xml_harvard']

df[opinions_columns].head()

df[opinions_columns].isnull().sum()

# Extract and clean the opinions from each respective column

cleaning_functions = {
    'opinion_plain_text': clean_plain,
    'opinion_html': clean_html,
    'opinion_html_with_citations': clean_html,
    'opinion_html_anon_2020': clean_html,
    'opinion_html_columbia': clean_html,
    'opinion_html_lawbox': clean_html,
    'opinion_xml_harvard': clean_xml
}

for column, clean_func in cleaning_functions.items():
    df.loc[:, column] = df[column].apply(lambda x: clean_func(x) if pd.notna(x) else x)

df[opinions_columns].head()

# Create one column for the opinion to be used for modeling

In accordance with the [documentation](https://www.courtlistener.com/help/api/rest/case-law/), the opinions used for modeling is selected based on the order below (from best to worst).

opinion_order = ["opinion_html_with_citations",
                 "opinion_html_columbia",
                 "opinion_html_lawbox",
                 "opinion_xml_harvard",
                 "opinion_html_anon_2020",
                 "opinion_html",
                 "opinion_plain_text"]

df[['opinion', 'opinion_source']] = df.apply(
    lambda row: pd.Series(get_first_non_null(row, opinion_order)), axis=1
)
df[opinions_columns + ['opinion', 'opinion_source']].head()

### Confirm no blank opinions

assert df["opinion"].isnull().sum() == 0

# Clean up the columns & save the cleaned sample

### Drop the separate opinion columns and keep only the combined opinion column

df = df.drop(columns=opinions_columns)
df.columns

### Add a column for the opinion word count

df["opinion_word_count"] = df["opinion"].astype(str).str.split().str.len()
df["opinion_word_count"].describe()

### Reorder the columns

columns = df.columns.tolist()

columns.remove('opinion')
columns.insert(1, 'opinion')

columns.remove('opinion_source')
columns.insert(2, 'opinion_source')

columns.remove('opinion_word_count')
columns.insert(3, 'opinion_word_count')

df = df[columns]
df.columns

### Save the cleaned file

df.to_csv("outputs/1b.opinions_cleaned.csv", index=False)
