==> notebooks_&_data/3a.check_similarities.ipynb <==
# Check similarities

This notebook outlines the steps I undertook to confirm the relevant queries are more similar to the opinion than the irrelevant queries are.

# Import libraries

Restart kernel after installing torch

#%pip install transformers -q
#%pip install torch -q

import numpy as np
import pandas as pd

from transformers import AutoTokenizer, AutoModel
from sklearn.feature_extraction.text import TfidfVectorizer

# Global variables

TOKENIZER = AutoTokenizer.from_pretrained("nlpaueb/legal-bert-base-uncased")
MODEL = AutoModel.from_pretrained("nlpaueb/legal-bert-base-uncased")
MAX_LENGTH = 512 #max_length for legal-bert model

# Helper functions

def calculate_cosine(string1, string2):
    similarity = np.dot(string1, string2)/np.sqrt(np.dot(string1, string1) * np.dot(string2, string2))
    return 0 if np.isnan(similarity) else similarity

def get_tfidf_similarity(query, result):
    vectorizer = TfidfVectorizer()
    model = vectorizer.fit([result])
    result_embedding = model.transform([result]).toarray().flatten()
    query_embedding = model.transform([query]).toarray().flatten()
    return calculate_cosine(query_embedding, result_embedding)
    
def get_llm_similarity(query, result):
    # get the CLS token representing the query sentence from the model
    inputs = TOKENIZER(query,
                       return_tensors="pt",
                       padding='max_length',
                       truncation=True, 
                       max_length=MAX_LENGTH)
    outputs = MODEL(**inputs)
    query_embedding = outputs.last_hidden_state[0][0].detach().numpy()

    # get the CLS token representing the result sentence from the model
    inputs = TOKENIZER(result,
                       return_tensors="pt",
                       padding='max_length',
                       truncation=True, 
                       max_length=MAX_LENGTH)
    outputs = MODEL(**inputs)
    result_embedding = outputs.last_hidden_state[0][0].detach().numpy()

    # calculate the cosine similarity between the two CLS tokens
    return calculate_cosine(query_embedding, result_embedding)

# Load the data

stmt = pd.read_csv("outputs/2a.queries_generated.csv")
qstn = pd.read_csv("outputs/2b.questions_generated.csv")

# Check query-result sets

## Calculate the cosine similarity based on TF-IDF matrix

stmt.loc[:, "relevant_tfidf"] = stmt.apply(lambda row: get_tfidf_similarity(row["relevant_query_stmt"], row["opinion"]), axis=1)
stmt.loc[:, "irrelevant_tfidf"] = stmt.apply(lambda row: get_tfidf_similarity(row["irrelevant_query_stmt"], row["opinion"]), axis=1)

stmt.head()

## Use Legal-Bert as double check

double_check = stmt[stmt["relevant_tfidf"] < stmt["irrelevant_tfidf"]]
len(double_check)

double_check.loc[:, "relevant_llm"] = double_check.apply(lambda row: get_llm_similarity(row["relevant_query_stmt"], row["opinion"]), axis=1)
double_check.loc[:, "irrelevant_llm"] = double_check.apply(lambda row: get_llm_similarity(row["irrelevant_query_stmt"], row["opinion"]), axis=1)

## Manual check

manual_check = double_check[double_check["relevant_llm"] < double_check["irrelevant_llm"]]
len(manual_check)

manual_review = manual_check.sample(5, random_state=42)[["opinion", "relevant_query_stmt", "irrelevant_query_stmt"]]
manual_review

manual_review.iloc[4]["relevant_query_stmt"]

manual_review.iloc[4]["irrelevant_query_stmt"]

manual_review.iloc[4]["opinion"]

# Check question-result sets

## Calculate the cosine similarity based on TF-IDF matrix

qstn.loc[:, "relevant_tfidf"] = qstn.apply(lambda row: get_tfidf_similarity(row["relevant_query_qstn"], row["opinion"]), axis=1)
qstn.loc[:, "irrelevant_tfidf"] = qstn.apply(lambda row: get_tfidf_similarity(row["irrelevant_query_qstn"], row["opinion"]), axis=1)

qstn.head()

## Use Legal-Bert as double check

double_check = qstn[qstn["relevant_tfidf"] < qstn["irrelevant_tfidf"]]
len(double_check)

double_check.loc[:, "relevant_llm"] = double_check.apply(lambda row: get_llm_similarity(row["relevant_query_qstn"], row["opinion"]), axis=1)
double_check.loc[:, "irrelevant_llm"] = double_check.apply(lambda row: get_llm_similarity(row["irrelevant_query_qstn"], row["opinion"]), axis=1)

## Manual check

manual_check = double_check[double_check["relevant_llm"] < double_check["irrelevant_llm"]]
len(manual_check)

manual_review = manual_check.sample(5, random_state=42)[["opinion", "relevant_query_qstn", "irrelevant_query_qstn"]]
manual_review

manual_review.iloc[4]["relevant_query_qstn"]

manual_review.iloc[4]["irrelevant_query_qstn"]

manual_review.iloc[4]["opinion"]
