==> notebooks_&_data/3b.combine_datasets.ipynb <==
# Combine the datasets

The notebook outlines the steps undertook to combine the datasets from previous steps to create a final dataset. Also split the dataset to train vs test set for downstream tasks.

# Import libraries

import numpy as np
import pandas as pd

from sklearn.model_selection import train_test_split

# Load the data

df = pd.read_csv("outputs/1b.opinions_cleaned.csv")
stmt = pd.read_csv("outputs/2a.queries_generated.csv")
qstn = pd.read_csv("outputs/2b.questions_generated.csv")

df.columns

stmt.columns

qstn.columns

# Combine the datasets

combined = df.merge(stmt[["opinion_id", "opinion_4omini_tokens", 'relevant_query_stmt', 'irrelevant_query_stmt']], how="left", on="opinion_id")
combined = combined.merge(qstn[["opinion_id", 'relevant_query_qstn', 'irrelevant_query_qstn']], how="left", on="opinion_id")
len(combined)

combined.columns

# Reorder the columns

columns = combined.columns.tolist()

columns.remove('opinion_4omini_tokens')
columns.insert(4, 'opinion_4omini_tokens')

combined = combined[columns]
combined.columns

# Save the combined & cleaned data

combined.to_csv("outputs/3.dataset.csv", index=False)

# Split the combined data to train vs test

train_df, test_df = train_test_split(combined, test_size=450, random_state=42)

len(train_df)

len(test_df)

train_df.to_csv("outputs/3.train.csv", index=False)
test_df.to_csv("outputs/3.test.csv", index=False)
