==> notebooks_&_data/6a.finetune_datasets.ipynb <==
# Generate dataset used for finetuning

This notebook outlines the steps undertook to generate the dataset used for finetuning. We already generated the train vs test datasets previously (and had not used the train dataset for any experiments so far), we want to ensure there is no dataleakage between the train and test datasets on opinion, cluster, and docket level.

# Import libraries

import pandas as pd

# Load the datasets

train = pd.read_csv("outputs/3.train.csv")
test = pd.read_csv("outputs/3.test.csv")

# Check for possible leakage

print(len(train[train["docket_id"].isin(test["docket_id"])]))
print(len(train[train["docket_number"].isin(test["docket_number"])]))
print(len(train[train["cluster_id"].isin(test["cluster_id"])]))
print(len(train[train["opinion_id"].isin(test["opinion_id"])]))

print(len(test[test["docket_id"].isin(train["docket_id"])]))
print(len(test[test["docket_number"].isin(train["docket_number"])]))
print(len(test[test["cluster_id"].isin(train["cluster_id"])]))
print(len(test[test["opinion_id"].isin(train["opinion_id"])]))

# Resolve leakage and generate validation set

val = train[train["docket_number"].isin(test["docket_number"])]
len(val)

train = train[~train["docket_number"].isin(test["docket_number"])]
len(train)

len(test)

In practice, we may want the test set to be of similar size to the validation set, but since we had already generated the test set for prior experiments, to ensure consistent metric measurements, we held the test set as is.

# Confirm no leakage

assert len(train[train["docket_id"].isin(test["docket_id"])]) == 0
assert len(train[train["docket_number"].isin(test["docket_number"])]) == 0
assert len(train[train["cluster_id"].isin(test["cluster_id"])]) == 0
assert len(train[train["opinion_id"].isin(test["opinion_id"])]) == 0

assert len(train[train["docket_id"].isin(val["docket_id"])]) == 0
assert len(train[train["docket_number"].isin(val["docket_number"])]) == 0
assert len(train[train["cluster_id"].isin(val["cluster_id"])]) == 0
assert len(train[train["opinion_id"].isin(val["opinion_id"])]) == 0

assert len(test[test["docket_id"].isin(train["docket_id"])]) == 0
assert len(test[test["docket_number"].isin(train["docket_number"])]) == 0
assert len(test[test["cluster_id"].isin(train["cluster_id"])]) == 0
assert len(test[test["opinion_id"].isin(train["opinion_id"])]) == 0

assert len(val[val["docket_id"].isin(train["docket_id"])]) == 0
assert len(val[val["docket_number"].isin(train["docket_number"])]) == 0
assert len(val[val["cluster_id"].isin(train["cluster_id"])]) == 0
assert len(val[val["opinion_id"].isin(train["opinion_id"])]) == 0

# Remove records that have fewer than 100 tokens (or ~50 words) as they do not have enough meaningful context for retrieval

train = train[train["opinion_word_count"] > 50]
len(train)

val = val[val["opinion_word_count"] > 50]
len(val)

test = test[test["opinion_word_count"] > 50]
len(test)

# Save the dataset for future use

train.to_csv("outputs/4.finetune_train.csv", index=False)
val.to_csv("outputs/4.finetune_val.csv", index=False)
test.to_csv("outputs/4.finetune_test.csv", index=False)
